+++ mmrazor/models/quantizers/__init__.py
@@ -1,10 +1,13 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .base import WithoutDeployQuantizer, WithDeployQuantizer
+from .base import BaseQuantizer
+from .academic_quantizer import AcademicQuantizer
+from .native_quantizer import NativeQuantizer
 from .openvino_quantizer import OpenVINOQuantizer
 from .tensorrt_quantizer import TensorRTQuantizer
 
 __all__ = [
-    'WithoutDeployQuantizer',
-    'WithDeployQuantizer',
+    'BaseQuantizer',
+    'AcademicQuantizer',
+    'NativeQuantizer',
     'TensorRTQuantizer', 
     'OpenVINOQuantizer']
