+++ mmrazor/models/task_modules/tracer/fx/__init__.py
@@ -1,8 +1,18 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .custom_tracer import (CustomTracer, UntracedMethodRegistry,
                             build_graphmodule, custom_symbolic_trace)
+from .graph_utils import (del_fakequant_before_module, 
+                          del_fakequant_after_module,
+                          del_fakequant_before_target,
+                          del_fakequant_after_target)
 
 __all__ = [
-    'CustomTracer', 'UntracedMethodRegistry', 'custom_symbolic_trace',
-    'build_graphmodule'
+    'CustomTracer', 
+    'UntracedMethodRegistry', 
+    'custom_symbolic_trace',
+    'build_graphmodule',
+    'del_fakequant_before_module',
+    'del_fakequant_after_module',
+    'del_fakequant_before_target',
+    'del_fakequant_after_target'
 ]
